move `get_morpheus_root` to test_morpheus.hpp
